skyFn_80089710 90.10->100%: type the flag temp u8 stateActive (matching the 1-bit 'active' field) and compare directly (stateActive != enabled, dropping the redundant (u8) cast) — the explicit cast created an Iro convert node that opt_loop_invariants hoisted into the unrolled loop's preheader, stealing r7 and cascading a full reg-perm; the u8 form fuses the promotion into the memory compare (per-iter clrlwi at ISEL, post-LICM) and evaluates active first, matching retail

+++ src/main/sky.c
@@ -706,7 +706,7 @@ void skyFn_80089710(int flags, u8 enabled, int startComplete)
 {
     u8* sky;
     u32 flagBit;
-    u32 stateActive;
+    u8 stateActive;
 
     sky = gSkyState;
     if (sky == NULL)
@@ -720,7 +720,7 @@ void skyFn_80089710(int flags, u8 enabled, int startComplete)
         {
             sky = gSkyState;
             stateActive = ((SkyBlendStateFlags*)(sky + flagBit * 0xa4 + 0xc1))->active;
-            if (stateActive != (u8)enabled)
+            if (stateActive != enabled)
             {
                 if (startComplete != 0)
                 {
